+++ src/App.js
@@ -3,10 +3,10 @@ import Index from "./page";
 import Search from "./page/search/search";
 import {BrowserRouter, Route, Routes} from "react-router-dom";
 import Layout from "./page/layout";
-import CourseAdd from "./page/admin/course/add/courseAdd";
+import CourseAdd from "./page/admin/management/courses/add/courseAdd";
 import AdminNav from "./page/admin/adminNav";
 import OfferableCourses from "./page/admin/course/offerableCourses";
-import CourseEdit from "./page/admin/course/edit/courseEdit";
+import CourseEdit from "./page/admin/management/courses/edit/courseEdit";
 import AdminDashboard from "./page/admin/adminDashboard";
 import UserPage from "./page/userPage";
 import About from "./page/about";
@@ -14,6 +14,8 @@ import Checkout from "./page/checkout";
 import Course from "./page/course";
 import Providers from "./page/admin/management/providers/providers";
 import Courses from "./page/admin/management/courses/courses";
+import ProviderAdd from "./page/admin/management/providers/add/providerAdd";
+import ProviderEdit from "./page/admin/management/providers/edit/providerEdit";
 
 function App() {
     return (
@@ -36,13 +38,13 @@ function App() {
                         <Route path={"/admin/management"}>
                             <Route path={"providers"}>
                                 <Route index element={<Providers/>}/>
-                                <Route path={"add"} element={<div>Add Provider</div>}/>
-                                <Route path={"edit/:id"} element={<div>Edit Provider</div>}/>
+                                <Route path={"add"} element={<ProviderAdd /> }/>
+                                <Route path={"edit/:id"} element={<ProviderEdit/>}/>
                             </Route>
                             <Route path={"courses"}>
                                 <Route index element={<Courses/>}/>
-                                <Route path={"add"} element={<div>Add Course</div>}/>
-                                <Route path={"edit/:id"} element={<div>Edit Course</div>}/>
+                                <Route path={"add"} element={<CourseAdd/> }/>
+                                <Route path={"edit/:id"} element={<CourseEdit/>}/>
                             </Route>
                             <Route path={"users"} element={<div>Users</div>}/>
                         </Route>
