+++ src/page/admin/management/courses/add/courseAdd.css
@@ -1,14 +1,8 @@
 
-
-
-button {
-    background-color: transparent;
-    border: none;
-}
-
-
-.page section {
-    margin: 2rem 0;
+.courseInfo-page {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
 }
 
 #course-info {
@@ -52,6 +46,7 @@ button {
     width: 100%;
     height: 2.5rem;
     margin: 1rem 1rem;
+
     box-sizing: border-box;
     border: 1px solid var(--shadow);
     font-size: 1rem;
@@ -67,7 +62,9 @@ button {
 }
 
 #course-info textarea {
-    height: 10rem;
+    height: 16rem;
+    padding: 1rem;
+    resize: none;
     /*padding-left: 2rem;*/
 }
