+++ src/page/admin/management/courses/add/courseAdd.jsx
@@ -1,43 +1,9 @@
+import "./courseAdd.css";
+import {Form} from "react-router-dom";
 
-import {Form, useParams} from "react-router-dom";
-import {useEffect, useState} from "react";
+export default function CourseAdd() {
 
-class courseEntity {
-    constructor(id, category, closestCourse, credits, description, diffLevel, hoursWeek, imgLink, relatedCert, title) {
-        this.id = id;
-        this.category = category;
-        this.closestCourse = closestCourse;
-        this.credits = credits;
-        this.description = description;
-        this.diffLevel = diffLevel;
-        this.hoursWeek = hoursWeek;
-        this.imgLink = imgLink;
-        this.relatedCert = relatedCert;
-        this.title = title;
-    }
-}
 
-export default function CourseEdit(courseId) {
-
-    // TODO: Once backend returns larger object with provider, price, currency etc. Update this to reflect that.
-
-    const { id } = useParams();
-
-    console.log(id)
-
-    const [loading, setLoading] = useState(true);
-    const [course, setCourse] = useState([]);
-
-    useEffect(() => {
-        console.log("fetching")
-        fetch("http://localhost:8080/api/course/" + id)
-            .then((r) => r.json())
-            .then((c) => {
-                console.log(c)
-                setCourse(new courseEntity(c.id, c.category, c.closestCourse, c.credits, c.description, c.diffLevel, c.hoursWeek, c.imgLink, c.relatedCert, c.title));
-                setLoading(false);
-            })
-    }, []);
 
     /**
      * Handle API call for uploading images.
@@ -78,7 +44,7 @@ export default function CourseEdit(courseId) {
         console.log(value);
 
         const requestOptions = {
-            method: 'PUT',
+            method: 'POST',
             headers: {'Content-Type': 'application/json'},
             body: JSON.stringify(value)
         };
@@ -112,47 +78,41 @@ export default function CourseEdit(courseId) {
         });
     }
 
-    if (loading) {
-        return <div>Loading...</div>
-    }
-
     return (
         <div className="courseInfo-page">
             <div>
-                <form onSubmit={handleSubmit} action={"http://localhost:3000/course/" + courseId} method="PUT">
+                <form onSubmit={handleSubmit} action="http://localhost:3000/course" method="POST">
                     <section id="course-info">
-                        <input id={"id"} name={"id"} type={"number"} hidden={true} value={course.id}/>
-
                         <div className="input-wrapper"><label htmlFor="course-name">Course Name</label>
-                            <input type="text" id="course-name" name="title" defaultValue={course.title} required/></div>
+                            <input type="text" id="course-name" name="title" required/></div>
 
                         <div className="input-wrapper"><label htmlFor="course-description">Course Description</label>
-                            <textarea id="course-description" name="description" defaultValue={course.description} required></textarea></div>
+                            <textarea id="course-description" name="description" required></textarea></div>
 
                         <div className="group-3">
                             <div className="input-wrapper">
                                 <label htmlFor="difficulty-level">Difficulty Level</label>
-                                <select name="diffLevel" id="difficulty-level" defaultValue={course.diffLevel} required>
+                                <select name="diffLevel" id="difficulty-level" required>
                                     <option value="0">Beginner</option>
                                     <option value="1">Intermediate</option>
                                     <option value="2">Expert</option>
                                 </select></div>
 
                             <div className="input-wrapper"><label htmlFor="course-credits">ECTS Credits</label>
-                                <input type="number" step={".5"} id="course-credits" name="credits" defaultValue={course.credits} required/></div>
+                                <input type="number" step={".5"} id="course-credits" name="credits" required/></div>
 
                             <div className="input-wrapper"><label htmlFor="course-duration">Duration</label>
-                                <input type="number" id="course-duration" name="hoursWeek" defaultValue={course.hoursWeek} required/></div>
+                                <input type="number" id="course-duration" name="hoursWeek" required/></div>
                         </div>
 
                         <div className="group-2">
                             <div className="input-wrapper"><label htmlFor="related-certification">Related
                                 Certification</label>
-                                <input type="text" id="related-certification" name="relatedCert" defaultValue={course.relatedCert} required/>
+                                <input type="text" id="related-certification" name="relatedCert" required/>
                             </div>
 
                             <div className="input-wrapper"><label htmlFor="course-category">Category</label>
-                                <select name="catergory" id="course-category" defaultValue={course.category} required>
+                                <select name="catergory" id="course-category" required>
                                     <option value="it">Information Technologies</option>
                                     <option value="dm">Digital Marketing</option>
                                     <option value="be">Business and Entrepreneurship</option>
@@ -160,35 +120,19 @@ export default function CourseEdit(courseId) {
                                 </select></div>
                         </div>
 
-                        <div className="group-2">
-                            <div className="input-wrapper"><label htmlFor="course-price">Price</label>
-                                <input type="number" id="course-price" name="price" defaultValue={course.price} required/></div>
 
-                            <div className="input-wrapper"><label htmlFor="price-currency">Currency</label>
-                                <select name="currency" id="price-currency" defaultValue={course.currency} required>
-                                    <option value="usd">USD</option>
-                                    <option value="eur">EUR</option>
-                                    <option value="nok">NOK</option>
-                                </select></div>
-                        </div>
 
-                        <div className="group-2">
-                            {/*TODO: Replace with calendar modal later (javascript component)*/}
-                            <div className="input-wrapper"><label htmlFor="course-date">Start Date</label>
-                                <input type="date" id="course-date" name="closestCourse" defaultValue={course.closestCourse} required/></div>
-
-                            {/*TODO: Add preview of uploaded image (javascript component)*/}
-                            <div className="input-wrapper">
-                                <label htmlFor="course-image">Course Image</label>
-                                <input type="file" id="course-image" name="imgLink" required/>
-                            </div>
+                        {/*TODO: Add preview of uploaded image (javascript component)*/}
+                        <div className="input-wrapper">
+                            <label htmlFor="course-image">Course Image</label>
+                            <input type="file" id="course-image" name="imgLink" required/>
                         </div>
 
                         <div className="input-wrapper"><label htmlFor="course-keywords">Keywords separated by
                             comma</label>
                             <input type="text" id="course-keywords" name="keywords" required/></div>
 
-                        <button className="cta-button" type="submit">Update Course</button>
+                        <button className="cta-button courseAdmin-button" type="submit">Add Course</button>
                     </section>
                 </form>
             </div>
