+++ src/page/admin/management/providers/add/providerAdd.css
@@ -1,18 +1,19 @@
 
-.courseInfo-page {
+.providerInfo-page {
     display: flex;
     flex-direction: column;
+    margin: 1rem 2rem;
     gap: 1rem;
 }
 
-#course-info {
+#provider-info {
     display: flex;
     flex-direction: column;
     gap: 1rem;
     margin: 2rem;
 }
 
-#course-info .input-wrapper {
+#provider-info .input-wrapper {
     position: relative;
     display: inline-block;
     margin: 1rem 0;
@@ -38,20 +39,21 @@
     flex: 1 1 calc(33.33% - 0.7rem); /* Three columns with gap */
 }
 
-#course-info input[type="text"],
-#course-info input[type="number"],
-#course-info input[type="date"],
-#course-info select,
-#course-info textarea {
+#provider-info input[type="text"],
+#provider-info input[type="number"],
+#provider-info input[type="date"],
+#provider-info select,
+#provider-info textarea {
     width: 100%;
     height: 2.5rem;
     margin: 1rem 1rem;
+
     box-sizing: border-box;
     border: 1px solid var(--shadow);
     font-size: 1rem;
 }
 
-#course-info input[type="file"] {
+#provider-info input[type="file"] {
     width: 100%;
     height: 2.5rem;
     margin: 1rem 1rem;
@@ -60,12 +62,14 @@
 
 }
 
-#course-info textarea {
-    height: 10rem;
+#provider-info textarea {
+    height: 16rem;
+    padding: 1rem;
+    resize: none;
     /*padding-left: 2rem;*/
 }
 
-#course-info .cta-button {
+#provider-info .cta-button {
     margin: 1rem 1rem;
 }
