+++ src/page/admin/management/providers/add/providerAdd.jsx
@@ -1,0 +1,35 @@
+import "./providerAdd.css"
+
+export default function ProviderAdd() {
+
+    return(
+        <div className="providerInfo-page">
+            <h1>Add Provider</h1>
+            <form className="providerInfo-form">
+                <section id="provider-info">
+                    <div className="input-wrapper"><label htmlFor="provider-name">Provider Name</label>
+                        <input type="text" id="provider-name" name="name" required/></div>
+
+
+                    <div className="group-2">
+                        {/*TODO: Add preview of uploaded image (javascript component)*/}
+                        <div className="input-wrapper">
+                            <label htmlFor="provider-image">Provider Image</label>
+                            <input type="file" id="provider-image" name="imgLink" required/>
+                        </div>
+
+                        {/*TODO: Add preview of uploaded image (javascript component)*/}
+                        <div className="input-wrapper">
+                            <label htmlFor="provider-alt-image">Alternative Provider Image</label>
+                            <input type="file" id="provider-alt-image" name="imgLinkAlt" required/>
+                        </div>
+                    </div>
+
+
+                    <button type="submit" className={"button cta-button"}>Add Provider</button>
+                </section>
+
+            </form>
+        </div>
+    )
+}
